+++ android/app/lib/screens/player/now_playing_screen.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:audio_service/audio_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 
 import '../../api/api_error.dart';
 import '../../models/seed_track.dart';